Add a basic integration/compatibility testing setup
This builds a container image that has a SSH server running as its
entrypoint. From there, we can run both Ansible and Sophons against it,
and export the resulting filesystem via `docker export`. This will then
yield two tar archives that we can easily compare to make sure they're
identical. If they're not, then we'll know we have built an
implementation that differs from Ansible.