Add rudimentary support for variables and Jinja
Inventory variables are now supported and can be injected into plays,
provided they're used only in strings. Support is not implemented for
parameters that are not strings (booleans, numbers). Since this depends
on a Go implementation of Jinja, there may be differences in the finer
details of Jinja pipelines, which may need addressing later on.