Add primitive support for roles
Right now this only implements tasks with no name, doesn't account for
roles variables or defaults yet, doesn't deal with handlers, files or
templates at all. It also assumes roles are found under a `roles`
directory present next to the playbook targetted, while Ansible has
support for more locations.

The implementation of roles loading is a bit funky: to keep the
execution local to every controlled node, the dialer will tar the `roles`
directory and copy that archive via SFTP (just like other content, for
instance the playbook). The executer will then unpack that archive at
runtime, parse it and execute the relevant code.

Doing so – keeping the execution local always – means we get to keep
really fast execution overall, and we avoid a ton of back and forth
between the controller and the controlled nodes.