Switch ansible.builtin.apt definition to protobuf
The long-term goal with this is to enable the use of gRPC (over SSH or
not) as a transport mechanism and to decompose the executer further. For
now let's move only one builtin to use protobuf as a starting point and
expand from there. This requires adding a custom unmarshaler to deal
with `name` being potentially a single string. We also need a custom
marshaler to deal with needing deep copies of tasks for loop processing.